+++ libs/support/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 
     <application>
         <activity
-            android:name="org.wordpress.android.support.feature.aibot.AIBotSupportActivity"
+            android:name="org.wordpress.android.support.feature.aibot.ui.AIBotSupportActivity"
             android:exported="false"
             android:theme="@style/Theme.Material3.DayNight.NoActionBar" />
     </application>
